backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation and serialization
pydantic==2.5.0
pydantic[email]==2.5.0

# AI/ML Libraries
torch>=2.0.0
torchaudio>=2.0.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# File handling and utilities
aiofiles==23.2.1

# ASGI server
gunicorn==21.2.0

# Optional: For better performance
python-multipart==0.0.6